requirements.txt
fastapi
uvicorn
python-multipart
pillow
aiofiles
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu